moved nortverse module to its own crate with some hasty patches
