+++ .gitignore
@@ -16,10 +16,9 @@
 
 /wordle/
 
-/macros/Cargo.lock
 /macros/target/
 
-/Cargo.lock
+*/Cargo.lock
 
 /config
