+++ nortverse/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "slimebot-nortverse"
+edition = "2021"
+readme = "README.md"
+license = "AGPL-3.0"
+
+[profile.release]
+strip = true
+lto = true
+codegen-units = 1
+opt-level = "z"
+
+[profile.dev]
+lto = false
+opt-level = 0
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+mongodb = "2.8.2"
+poise = "0.6.1"
+reqwest = "0.12.12"
+scraper = "0.22.0"
+serde = { version = "1.0.217", features = ["derive"] }
+thiserror = "2.0.10"
+tokio = { version = "1.43.0", features = ["sync"] }
+tracing = "0.1.41"
+tracing-unwrap = "1.0.1"
+thisslime = "0.5.1"
+rand = "0.8.5"
+url = "2.5.4"
