+++ src/errors.rs
@@ -92,7 +92,6 @@ pub enum CommandError {
     #[error("error from minecraft api: {0}")]
     Minecraft(#[from] crate::commands::minecraft::Error),
 
-    #[cfg(feature = "nortverse")]
     #[error("error from nortverse api: {0}")]
     Nortverse(#[from] crate::commands::nortverse::Error),
